market-ai/marketing_ai_employees_growing_brand_bundle/requirements.txt
openai>=1.51.0
openai-agents>=0.2.0
python-dotenv>=1.0.1
pydantic>=2.8.2
